Refactor OpenY upgrade tool
There is a hard rule which can't be skipped, and should be rethinked from the foundation
OpenY upgrade tool stops config updates when there is a manual change detected.
Which is wrong
User story
I'm an OpenY developer and I need to upgrade from 8.0.1 to 8.1.12
For that, because of large number of updates and impossibility to predict quality of upgrade path support for the contrib modules it is impossible to upgrade in one leap - from 8.0.1 to 8.1.12. 
For that I have to update from version to version, relying on some investigation from the OpenY community https://github.com/ymcatwincities/openy/wiki/%5BDraft%5D-OpenY-upgrade-how-to-for-Developers
Thus, after first detected manually changed config I'm blocked from upgradig to latest version of OpenY with a huge scope in front of me
As a better resolution I'd like to be able to update to latest version even if my configs and manual changes are overridden. 
And as a feature from OpenY upgrade tool would be cool to keep is to show the diff between latest config from updated OpenY version of my site and old config that was overridden during upgrade